+++ src/components/panes/WorkbenchPane.vue
@@ -14,7 +14,7 @@
 
     <div class="workbench-pane__objectives">
       <div v-for="objective in workbenchObjectives" :key="objective.id" class="">
-        <okr-link-card
+        <objective-link-card
           :route="{ name: 'ObjectiveHome', params: { objectiveId: objective.id } }"
           :title="objective.name"
           :objective-id="objective.id"
@@ -40,7 +40,7 @@
 import { mapActions, mapGetters, mapState } from 'vuex';
 import { PktButton } from '@oslokommune/punkt-vue2';
 import PaneWrapper from '@/components/panes/PaneWrapper.vue';
-import OkrLinkCard from '@/components/OkrLinkCard.vue';
+import ObjectiveLinkCard from '@/components/ObjectiveLinkCard.vue';
 import ProgressBar from '@/components/ProgressBar.vue';
 
 export default {
@@ -48,7 +48,7 @@ export default {
 
   components: {
     PaneWrapper,
-    OkrLinkCard,
+    ObjectiveLinkCard,
     ProgressBar,
     PktButton,
   },
